implemented authentication

+++ src/main/java/com/r0r5chach/services/SingleDocumentService.java
@@ -1,5 +1,6 @@
 package com.r0r5chach.services;
 
+import static com.r0r5chach.services.AuthService.tokenExists;
 import static org.bson.Document.parse;
 
 import org.bson.Document;
@@ -18,17 +19,23 @@ public class SingleDocumentService extends DBService {
      * @return A string containing the response in JSON format
      */
     public static String get(Request req, Response res) {
-        res.type("application/json");
-        Document query = parse(req.body());
-        MongoCollection<Document> col = client.getCollection(req.queryMap().get("collection").value());
-        
-        if (col.countDocuments(query) > 0) {
-            res.status(200);
-            return col.find(query).first().toJson();
-        } 
+        if (tokenExists(req.queryMap().get("token").value())) {
+            res.type("application/json");
+            Document query = parse(req.body());
+            MongoCollection<Document> col = client.getCollection(req.queryMap().get("collection").value());
+            
+            if (col.countDocuments(query) > 0) {
+                res.status(200);
+                return col.find(query).first().toJson();
+            } 
+            else {
+                res.status(500);
+                return "{\"response\":\"No Documents found\"}";
+            }
+        }
         else {
-            res.status(500);
-            return "{\"response\":\"No Documents found\"}";
+            res.status(401);
+            return "{\"response\":\"Unauthorized\"}";
         }
     }
 
@@ -40,18 +47,24 @@ public class SingleDocumentService extends DBService {
      * @return A string containing the response in JSON format
      */
     public static String post(Request req, Response res) {
-        Document item = parse(req.body());
+        if (tokenExists(req.queryMap().get("token").value())) {
+            Document item = parse(req.body());
 
-        try {
-            client.getCollection(req.queryMap().get("collection").value()).insertOne(item);
-        } 
-        catch (Exception e) {
-            res.status(500);
-            return "{\"response\":\"Insert failed\"}";
+            try {
+                client.getCollection(req.queryMap().get("collection").value()).insertOne(item);
+            } 
+            catch (Exception e) {
+                res.status(500);
+                return "{\"response\":\"Insert failed\"}";
+            }
+
+            res.status(201);
+            return "{\"response\":\"Insert successful\"}";
         }
-
-        res.status(201);
-        return "{\"response\":\"Insert successful\"}";
+        else {
+            res.status(401);
+            return "{\"response\":\"Unauthorized\"}";
+        }
     }
 
     /**
@@ -62,20 +75,26 @@ public class SingleDocumentService extends DBService {
      * @return A string containing the response in JSON format
      */
     public static String put(Request req, Response res) {
-        Document request = parse(req.body());
-        Document query = parse(request.getString("query"));
-        Document update = new Document().append("$set", parse(request.getString("update")));
+        if (tokenExists(req.queryMap().get("token").value())) {
+            Document request = parse(req.body());
+            Document query = parse(request.getString("query"));
+            Document update = new Document().append("$set", parse(request.getString("update")));
 
-        try {
-            client.getCollection(req.queryMap().get("collection").value()).updateOne(query, update);
-        } 
-        catch (Exception e) {
-            res.status(500);
-            return "{\"response\":\"Update failed\"}";            
+            try {
+                client.getCollection(req.queryMap().get("collection").value()).updateOne(query, update);
+            } 
+            catch (Exception e) {
+                res.status(500);
+                return "{\"response\":\"Update failed\"}";            
+            }
+
+            res.status(200);
+            return "{\"response\":\"Update successful\"}";
         }
-
-        res.status(200);
-        return "{\"response\":\"Update successful\"}";
+        else {
+            res.status(401);
+            return "{\"response\":\"Unauthorized\"}";
+        }
     }
 
     /**
@@ -86,17 +105,23 @@ public class SingleDocumentService extends DBService {
      * @return A string containing the response in JSON format
      */
     public static String delete(Request req, Response res) {
-        Document query = parse(req.body());
-        try {
-            client.getCollection(req.queryMap().get("collection").value()).deleteOne(query);
-        }
-        catch (Exception e) {
-            res.status(500);
-            return "{\"response\":\"Delete failed\"}";
-        }
+        if (tokenExists(req.queryMap().get("token").value())) {
+            Document query = parse(req.body());
+            try {
+                client.getCollection(req.queryMap().get("collection").value()).deleteOne(query);
+            }
+            catch (Exception e) {
+                res.status(500);
+                return "{\"response\":\"Delete failed\"}";
+            }
 
-        res.status(200);
-        return "{\"response\":\"Delete successful\"}";
+            res.status(200);
+            return "{\"response\":\"Delete successful\"}";
+        }
+        else {
+            res.status(401);
+            return "{\"response\":\"Unauthorized\"}";
+        }
     }
 
     public static String options(Request req, Response res) {
